refactor(benchmark): rename benchmark tool from stress to benchmark
